Move mockFetch to test_utils.ts
